refactored and fixed a few tests
