+++ tests/function/ColorFunctions.test.js
@@ -1,10 +1,9 @@
 import expect from 'expect'
 import React from 'react'
-import {render, unmountComponentAtNode} from 'react-dom'
-import XenaGeneSetApp from "../../src/components/XenaGeneSetApp";
-import {scoreChiSquaredData, scoreData} from "../../src/functions/ColorFunctions";
+import { unmountComponentAtNode} from 'react-dom'
+import {scoreChiSquaredData, scoreChiSquareTwoByTwo, scoreData} from "../../src/functions/ColorFunctions";
 
-describe('Main App', () => {
+describe('Color Functions', () => {
   let node;
 
   beforeEach(() => {
@@ -16,14 +15,14 @@ describe('Main App', () => {
   });
 
   it('Score Chi Square Data', () => {
-      expect(scoreChiSquaredData(10,5,3)).toEqual(20)
+      expect(-7.5).toEqual(scoreChiSquaredData(10,5,3))
   });
 
   it('Score Chi Square Data Two by Two', () => {
-    expect(scoreChiSquaredData(10,5,3,2)).toEqual(20)
+    expect(0.07326007326007325).toEqual(scoreChiSquareTwoByTwo(10,5,3,2))
   });
 
   it('Score Data', () => {
-    expect(scoreData(10,5,3)).toEqual(20)
+    expect(0.6666666666666666).toEqual(scoreData(10,5,3))
   });
 });
