+++ tests/function/DataFunctions.test.js
@@ -2,14 +2,18 @@ import expect from 'expect'
 import React from 'react'
 import { unmountComponentAtNode} from 'react-dom'
 import {
-  associateData,
+  addIndepProb,
+  associateData, createEmptyArray,
   getCopyNumberHigh,
   getCopyNumberLow,
   getCopyNumberValue, getGenePathwayLookup,
   getMutationScore, pruneColumns
 } from "../../src/functions/DataFunctions";
+import {times} from "underscore";
 
-describe('Main App', () => {
+
+
+describe('Data Functions', () => {
   let node;
 
   beforeEach(() => {
@@ -81,6 +85,32 @@ describe('Main App', () => {
   });
 
 
+  it('Create a simple array', () => {
+    let returnArray = createEmptyArray(20,5);
+    expect(returnArray.length).toEqual(20);
+    expect(returnArray[0].length).toEqual(5);
+    // expect(returnArray[5][3]).toEqual({total:0,mutation:0,cnv:0});
+    expect(returnArray[5][3]).toEqual({total:0,mutation4:0,mutation3:0,mutation2:0,mutation:0,cnv:0,cnvLow:0,cnvHigh:0});
+    returnArray[5][3] = {total:7,mutation:3,cnv:1};
+    expect(returnArray[5][3]).toEqual({total:7,mutation:3,cnv:1});
+    returnArray = new Array(20).fill(0).map(() => new Array(5).fill({total:0,mutation:0,cnv:0}));
+    expect(returnArray[5][3]).toEqual({total:0,mutation:0,cnv:0});
+
+  });
+
+  it('Calculates single function properly', () => {
+    // let returnArray = new Array(20).fill([]).map(() => new Array(5).fill({total:0,mutation:0,cnv:0}));
+    // let returnArray = times(20,fill([]).map(() => new Array(5).fill({total:0,mutation:0,cnv:0}));
+    let returnArray = times(20, () => times(5, () => DEFAULT_DATA_VALUE));
+    expect(returnArray.length===20);
+    expect(returnArray[0].length===5);
+    expect(returnArray[5][3]).toEqual({total:0,mutation4:0,mutation3:0,mutation2:0,mutation:0,cnv:0,cnvLow:0,cnvHigh:0});
+    returnArray[5][3] = {total:7,mutation:3,cnv:1};
+    expect(returnArray[5][3]).toEqual({total:7,mutation:3,cnv:1});
+    returnArray = new Array(20).fill(0).map(() => new Array(5).fill({total:0,mutation:0,cnv:0}));
+    expect(returnArray[5][3]).toEqual({total:0,mutation:0,cnv:0});
+
+  });
 
 
 });
